+++ hvisor/src/percpu.rs
@@ -2,13 +2,13 @@ use crate::arch::riscv::cpu::ArchCpu;
 use crate::consts::{INVALID_ADDRESS, PAGE_SIZE, PER_CPU_ARRAY_PTR, PER_CPU_SIZE};
 use crate::memory::addr::VirtAddr;
 pub struct PerCpu {
-    pub id: u64,
-    pub cpu_on_entry: u64,
+    pub id: usize,
+    pub cpu_on_entry: usize,
     pub arch_cpu: ArchCpu,
 }
 
 impl PerCpu {
-    pub fn new(cpu_id: u64) -> Self {
+    pub fn new(cpu_id: usize) -> Self {
         PerCpu {
             id: cpu_id,
             cpu_on_entry: INVALID_ADDRESS,
